Extract generic web deploy rollback dispatch
